Day 10, part 1: Change core trait to be u64 instead of u32.

+++ rust/twenty-twenty/src/core.rs
@@ -4,10 +4,10 @@ use std::path::PathBuf;
 use std::fs;
 
 pub trait AdventOfCodeSolver {
-    fn partOne(&self) -> u32 {
+    fn partOne(&self) -> u64 {
         return 0;
     }
-    fn partTwo(&self) -> u32 {
+    fn partTwo(&self) -> u64 {
         return 0;
     }
     fn day(&self) -> &str;
